Colortomata/requirements.txt
numpy~=2.2.3
pillow~=10.4.0
Flask~=3.1.0
waitress~=3.0.2